Change block and block results query calls to Tendermint to use HTTP instead of WebSocket, Fix duplicate message processing by removing cache when message is going to be process from handle message from message queue function, Remove unnecessary block results query call to Tendermint

+++ src/core/as.js
@@ -397,11 +397,12 @@ export async function handleMessageFromQueue(messageStr) {
       if (tendermint.latestBlockHeight <= message.height) {
         delete requestIdLocks[message.request_id];
         return;
+      } else {
+        await db.removeRequestReceivedFromMQ(requestId);
       }
     }
 
     await processRequest(message);
-    await db.removeRequestReceivedFromMQ(requestId);
     delete requestIdLocks[message.request_id];
   } catch (error) {
     const err = new CustomError({
@@ -454,6 +455,7 @@ export async function handleTendermintNewBlockHeaderEvent(
       requestIdsInTendermintBlock.map(async (requestId) => {
         if (requestIdLocks[requestId]) return;
         const request = await db.getRequestReceivedFromMQ(requestId);
+        if (request == null) return;
         await processRequest(request);
         await db.removeRequestReceivedFromMQ(requestId);
       })
